Fix level modun

+++ src/Libraries/LW.Services/LevelServices/LevelService.cs
@@ -1,4 +1,5 @@
-﻿using AutoMapper;
+﻿using System.Text;
+using AutoMapper;
 using LW.Data.Entities;
 using LW.Data.Repositories.LevelRepositories;
 using Serilog;
@@ -23,6 +24,7 @@ public class LevelService : ILevelService
     public async Task<ApiResult<bool>> Create(LevelDtoForCreate model)
     {
         var level = _mapper.Map<Level>(model);
+        level.KeyWord = level.Title.ToLower();
         await _levelRepository.CreateLevel(level);
         return new ApiResult<bool>(true, "Create level successfully");
     }
@@ -30,7 +32,15 @@ public class LevelService : ILevelService
 
     public async Task<ApiResult<bool>> Update(LevelDtoForUpdate model)
     {
-        var level = _mapper.Map<Level>(model);
+        var levelInDb = await _levelRepository.GetLevelById(model.Id);
+        if (levelInDb == null)
+        {
+            return new ApiResult<bool>(false, "Not found");
+        }
+        var obj = _mapper.Map(model, levelInDb);
+        // Sau khi ánh xạ, levelIbDb sẽ có các giá trị từ model:
+        var level = _mapper.Map<Level>(obj);
+        level.KeyWord = level.Title.ToLower();
         await _levelRepository.UpdateLevel(level);
         return new ApiResult<bool>(true, "Update level successfully");
     }
@@ -40,7 +50,14 @@ public class LevelService : ILevelService
         {
             return new ApiResult<bool>(true, "Id is invalid");
         }
-        await _levelRepository.UpdateStatusLevel(id);
+
+        var objLevel = await _levelRepository.GetLevelById(id);
+        if (objLevel == null)
+        {
+            return new ApiResult<bool>(true, "Not found !");
+        }
+        objLevel.IsActive = !objLevel.IsActive;
+        await _levelRepository.UpdateLevel(objLevel);
         return new ApiResult<bool>(true, "Update Status of level successfully");
     }
 
@@ -48,9 +65,19 @@ public class LevelService : ILevelService
     {
         if (id == null || id <= 0)
         {
-            return new ApiResult<bool>(true, "Id is invalid");
+            return new ApiResult<bool>(false, "Id is invalid");
         }
-        await _levelRepository.DeleteLevel(id);
+
+        var level = await _levelRepository.GetLevelById(id);
+        if (level == null)
+        {
+            return new ApiResult<bool>(false, "Not found !");
+        }
+        var isDeleted = await _levelRepository.DeleteLevel(id);
+        if (!isDeleted)
+        {
+            return new ApiResult<bool>(false, "Delete level failed");
+        }
         return new ApiResult<bool>(true, "Delete level successfully");
     }
